#45: move DictionaryGuidGenerator to DictionaryGuidGenerator.cs
